ajuste requirement 36
