+++ src/pages/TelaDetalheComida/index.js
@@ -7,8 +7,9 @@ import whiteHeartIcon from '../../images/whiteHeartIcon.svg';
 import { RecomendationCardMeal } from '../../components';
 import {
   getSpecificMealById,
-  getRecommendatedMeals,
+  // getRecommendatedMeals,
 } from '../../store/ducks/getDetailedMeal/actions';
+import { getRecommendatedDrinks } from '../../store/ducks/getDetailedDrink/actions';
 
 class TelaDetalheComida extends Component {
   async componentDidMount() {
@@ -17,8 +18,9 @@ class TelaDetalheComida extends Component {
         params: { id },
       },
     } = this.props;
-    const { getDetailedMealDispatch, getRecommendationMeals } = this.props;
-    await getRecommendationMeals();
+    const { getDetailedMealDispatch, getRecommendationDrinks } = this.props;
+    // await getRecommendationMeals();
+    await getRecommendationDrinks();
     await getDetailedMealDispatch(id);
   }
 
@@ -42,7 +44,8 @@ class TelaDetalheComida extends Component {
     const ingredientsArray = this.handleIngredients(meal);
     const measuresArray = this.handleMeasure(meal);
     const six = 6;
-    const { mealsRecommendStore } = this.props;
+    // const { mealsRecommendStore } = this.props;
+    const { drinksRecommendStore } = this.props;
     const settings = {
       dots: true,
       infinite: true,
@@ -50,6 +53,7 @@ class TelaDetalheComida extends Component {
       slidesToShow: 2,
       slidesToScroll: 2,
     };
+    // console.log(mealsRecommendStore);
     return (
       <>
         <img
@@ -93,14 +97,13 @@ class TelaDetalheComida extends Component {
         </div>
         <div data-testid="video">{meal.strYoutube}</div>
         <Slider { ...settings }>
-          {mealsRecommendStore.map((element, index) => {
+          {drinksRecommendStore.map((element, index) => {
             if (index < six) {
               return (
                 <RecomendationCardMeal
                   key={ element }
-                  data-testid={ `${index}-recomendation-card` }
-                  mealRecommended={ element }
-                  mealIndex={ index }
+                  drinkRecommended={ element }
+                  drinkIndex={ index }
                 />
               );
             }
@@ -131,18 +134,22 @@ TelaDetalheComida.propTypes = {
     }).isRequired,
   }).isRequired,
   getDetailedMealDispatch: PropTypes.func.isRequired,
-  mealsRecommendStore: PropTypes.arrayOf(PropTypes.Object).isRequired,
-  getRecommendationMeals: PropTypes.arrayOf(PropTypes.Object).isRequired,
+  drinksRecommendStore: PropTypes.arrayOf(PropTypes.Object).isRequired,
+  getRecommendationDrinks: PropTypes.func.isRequired,
+  // mealsRecommendStore: PropTypes.arrayOf(PropTypes.Object).isRequired,
+  // getRecommendationMeals: PropTypes.arrayOf(PropTypes.Object).isRequired,
 };
 
 const mapStateToProps = (state) => ({
   meal: state.detalhesDaReceitaComida.meal.meals,
-  mealsRecommendStore: state.detalhesDaReceitaComida.mealRecommend.meals,
+  drinksRecommendStore: state.detalhesDaReceitaBebida.drinkRecommend.drinks,
+  // mealsRecommendStore: state.detalhesDaReceitaComida.mealRecommend.meals,
 });
 
 const mapDispatchToProps = (dispatch) => ({
   getDetailedMealDispatch: (id) => dispatch(getSpecificMealById(id)),
-  getRecommendationMeals: () => dispatch(getRecommendatedMeals()),
+  getRecommendationDrinks: () => dispatch(getRecommendatedDrinks()),
+  // getRecommendationMeals: () => dispatch(getRecommendatedMeals()),
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(TelaDetalheComida);
